requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.11.5
colorama==0.4.1
configparser==3.5.1
cryptography==2.5
future==0.17.1
icdiff==1.9.1
Jinja2==2.10
jsnapy==1.3.2
junos-eznc==2.2.0
lxml==4.3.0
MarkupSafe==1.1.0
ncclient==0.6.3
netaddr==0.7.19
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.3.1
pyserial==3.4
PyYAML==3.13
scp==0.13.0
selectors2==2.0.1
six==1.12.0